--- a/src/Arrays.h
+++ b/src/Arrays.h
@@ -82,6 +82,16 @@ struct maskinfo
 
 };
 
+struct RiverInfo
+{
+	int nbir;
+	int nburmax; // size of (max number of) unique block with rivers  
+	int nribmax; // size of (max number of) rivers in one block
+	int* Xbidir; // array of block id for each river size()
+	int* Xridib; // array of river id in each block
+
+};
+
 
 // outzone info used to actually write the nc files (one nc file by zone, the default zone is the full domain)
 struct outzoneB