# Improve efficiency if may rivers are involved
I'm changing the structure of the model workflow to be efficient if a lot of rivers are in the model.
Currently kernels are launched in serial for each river. This is very inefficient when you have many rivers.
In an example using a crud model of the lower Waitaki River. 

| n rivers | Run times |
| -------- | ------- |
| 1 | 6    |
| 10 | 7    |
| 100    | 14    |
| 675 | 71 | 

This doesn't look too bad but things gets worse in other setting (may need a better example to make the point about the point of this)