@CyprienBosserelle 2024-10-17
using the same test:
675 river: 19s runtime!