@CyprienBosserelle 2024-10-17
Using the same setup as above but rerunning each scenarios. (Waitaki domain, single level (30m res). 3600 s run )



| n rivers | Run times old code |Run times new code|
| -------- | ------- |------- |
| 1 | 5  | 5 |
| 10 | 5 | 6|
| 100    | 14  | 7| 
| 675 | 73| 9-10 |

Results between old and new code are exactly identical but the new code is 7 times faster with 675 rivers. 

>[!TIP]
>SUCCESS