@CyprienBosserelle 2024-10-18
# Two Birds, one PR
## Momentum conserving rain input
Also added a momentum preserving correction for u and v when adding rain input.