@CyprienBosserelle 2025-03-11
# still to do

1. Automated bnd segment polygon using nc file bounding box
2. Keyword for Ocean side bnd
3. add param keywords for `bndrelaxtime` and `bndfiltertime`